changed double to float

+++ src/Oscillator/Oscillator.h
@@ -13,7 +13,7 @@ class Oscillator {
 
     public:
         Oscillator(int, WAVEFORM);
-        double getSample();
+        float getSample();
 };
 
 #endif
